fix: standardize route name using kebab-casing, rename publication folders and filenames
