+++ src/assets/static/images/publications/index.ts
