+++ src/i18n/en/index.ts
@@ -13,7 +13,7 @@ import { JobListing } from './JobListing'
 import { navbar } from './Navbar'
 import { OurPeople1 } from './OurPeople1'
 import { OurPeople2 } from './OurPeople2'
-import { Publications } from './OurPublications'
+import { Publications } from './publications'
 import { ourResearch } from './OurResearch'
 import { OurSupporters } from './OurSupporters'
 import { ourTechnology } from './OurTecnology'
