+++ src/i18n/es/index.ts
@@ -4,7 +4,7 @@ import { AboutUs } from './AboutUs'
 import { GetInvolved_ } from './GetInvolved'
 import { dashboard } from './Dashboard'
 import { JobDescription } from './JobDescription'
-import { Publications } from './OurPublications'
+import { Publications } from './publications'
 import { ourTechnology } from './OurTecnology'
 import { howItWorks } from './HowItWorks'
 import { OurPeople1 } from './OurPeople1'
